+++ src/main/java/com/fifteen/auction/domain/auction/infrastructure/RabbitBidRequestListener.java
@@ -1,7 +1,7 @@
 package com.fifteen.auction.domain.auction.infrastructure;
 
 import com.fifteen.auction.domain.auction.dto.event.BidRequestEvent;
-import com.fifteen.auction.domain.auction.service.BidService;
+import com.fifteen.auction.domain.auction.service.port.in.BidEventUseCase;
 import com.rabbitmq.client.Channel;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -15,18 +15,17 @@ import java.io.IOException;
 @RequiredArgsConstructor
 public class RabbitBidRequestListener {
 
-    private final BidService bidService;
+    private final BidEventUseCase bidEventUseCase;
 
     @RabbitListener(queues = "#{rabbitConfig.queueNames}")
     public void bidRequestListener(BidRequestEvent e, Channel channel, Message raw) throws IOException {
         try {
             log.info("[입찰 큐 이벤트 처리] auctionSeq={}\tbidderId={}\tbidPrice={}", e.getAuctionSeq(), e.getUserId(), e.getBidPrice());
-            bidService.handleBidFromQueue(e.getAuctionSeq(), e.getUserId(), e.getBidPrice());
+            bidEventUseCase.handleBidFromQueue(e.getAuctionSeq(), e.getUserId(), e.getBidPrice());
             channel.basicAck(raw.getMessageProperties().getDeliveryTag(), false);
         } catch (Exception ex) {
             channel.basicNack(raw.getMessageProperties().getDeliveryTag(), false, false);
             throw ex;
         }
     }
-
 }
