+++ src/main/java/com/fifteen/auction/domain/auction/service/BidEventService.java
@@ -2,23 +2,34 @@ package com.fifteen.auction.domain.auction.service;
 
 import com.fifteen.auction.domain.auction.dto.event.BidProcessEvent;
 import com.fifteen.auction.domain.auction.entity.Auction;
+import com.fifteen.auction.domain.auction.entity.Bid;
 import com.fifteen.auction.domain.auction.repository.auction.AuctionRepository;
+import com.fifteen.auction.domain.auction.repository.bid.BidRepository;
+import com.fifteen.auction.domain.auction.service.port.in.BidEventUseCase;
+import com.fifteen.auction.domain.auction.util.ClockHolder;
 import com.fifteen.auction.global.dto.error.ErrorCode;
 import com.fifteen.auction.global.dto.exception.ClientException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
+import java.time.LocalDateTime;
+
 @Component
 @RequiredArgsConstructor
-public class BidWorkerService {
+public class BidEventService implements BidEventUseCase {
 
     private final AuctionCacheService auctionCacheService;
 
     private final AuctionRepository auctionRepository;
+    private final BidRepository bidRepository;
+
+    private final ClockHolder clockHolder;
+    private final ApplicationEventPublisher applicationEventPublisher;
 
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
     @Transactional(propagation = Propagation.REQUIRES_NEW)
@@ -32,5 +43,27 @@ public class BidWorkerService {
         // 경매 표시가 반영
         auctionCacheService
                 .addNewHighPrice(event.getAuctionSeq(), event.getBidderId(), event.getBidPrice());
+    }
+
+    @Override
+    @Transactional
+    public void handleBidFromQueue(String auctionSeq, Long bidderId, Long bidPrice) {
+        Auction auc = auctionRepository.findOpenOneBySeqWithSeller(auctionSeq)
+                .orElseThrow(() -> new ClientException(ErrorCode.AUCTION_NOT_FOUND));
+
+        LocalDateTime bidAt = clockHolder.now();
+
+        if (bidAt.isAfter(auc.getExpiresAt())) {
+            throw new ClientException(ErrorCode.INVALID_BUY_NOW_REQUEST);
+        }
+
+        // bid price cache 체크
+        if (auctionCacheService.isBidUnderPrice(auc.getAuctionSeq(), bidPrice, auc.getBidUnit())) {
+            throw new ClientException(ErrorCode.LOW_BID_PRICE);
+        }
+
+        bidRepository.save(new Bid(auc, bidderId, bidPrice, bidAt));
+        applicationEventPublisher.publishEvent(
+                new BidProcessEvent(auctionSeq, bidderId, bidPrice, bidAt));
     }
 }
